+++ CHANGELOG.md
@@ -11,6 +11,12 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v3.74.4 (2022-01-14)
+
+#### :bug: Bug Fix
+
+* Fix bug with onAbort calling after aborting promise `core/promise/abortable`
+
 ## v3.74.3 (2022-01-13)
 
 #### :boom: Breaking Change